[refactor] also share load_driver between plugins
