+++ lib/logstash/inputs/jdbc.rb
@@ -1,6 +1,7 @@
 # encoding: utf-8
 require "logstash/inputs/base"
 require "logstash/namespace"
+require "logstash/plugin_mixins/jdbc/common"
 require "logstash/plugin_mixins/jdbc/jdbc"
 
 # this require_relative returns early unless the JRuby version is between 9.2.0.0 and 9.2.8.0
