+++ lib/logstash/plugin_mixins/jdbc/common.rb
@@ -2,6 +2,30 @@
 module LogStash module PluginMixins module Jdbc
   module Common
 
+    private
+
+    def load_driver
+      return @driver_impl if @driver_impl ||= nil
+
+      require "java"
+      require "sequel"
+      require "sequel/adapters/jdbc"
+
+      load_driver_jars
+      begin
+        @driver_impl = Sequel::JDBC.load_driver(@jdbc_driver_class)
+      rescue Sequel::AdapterNotFound => e # Sequel::AdapterNotFound, "#{@jdbc_driver_class} not loaded"
+        # fix this !!!
+        message = if jdbc_driver_library_set?
+                    "Are you sure you've included the correct jdbc driver in :jdbc_driver_library?"
+                  else
+                    ":jdbc_driver_library is not set, are you sure you included " +
+                        "the proper driver client libraries in your classpath?"
+                  end
+        raise LogStash::PluginLoadingError, "#{e}. #{message}"
+      end
+    end
+
     def load_driver_jars
       if jdbc_driver_library_set?
         @jdbc_driver_library.split(",").each do |driver_jar|
